If mysqld fails to start, don't endlessly loop
When mysqld fails to start it never writes out a pid file, so
wait_for_setup() calls waitpid() which reaps the child, and
then dies.

The destroy block then runs, and we call wait_for_stop(), where
waitpid will only ever see a '-1', and we'd loop forever.

Instead, only continue to waitpid on a running process if waitpid
returns 0 which means there are still children to be reaped.